added register page
